fix(docker-tests): eliminate warnings by reorganizing code
- Remove Secp256k1Secret re-export from env.rs; callers import from crypto
- Move funding locks (MYCOIN_LOCK, MYCOIN1_LOCK, FORSLP_LOCK, QTUM_LOCK)
  and get_funding_lock() from docker_ops.rs to utxo.rs
- Delete unused ZCOIN_GEN_TX_LOCK* locks (dead code)
- Simplify MYCOIN/MYCOIN1 constant cfg gates (ordermatch/watchers use
  string literals, not the constants)
- Remove docker-tests-zcoin from utxo module gate

All docker test features now compile with 0 warnings.
