+++ mm2src/mm2_main/tests/docker_tests/docker_ordermatch_tests.rs
@@ -4,7 +4,7 @@ use crate::integration_tests_common::enable_native;
 use common::block_on;
 use mm2_number::BigDecimal;
 use mm2_rpc::data::legacy::OrderbookResponse;
-use mm2_test_helpers::for_tests::{mm_dump, mycoin1_conf, mycoin_conf, MarketMakerIt};
+use mm2_test_helpers::for_tests::{mm_dump, MarketMakerIt};
 
 use mm2_test_helpers::structs::{
     BestOrdersResponse, BestOrdersV2Response, BuyOrSellRpcResult, MyOrdersRpcResult, OrderbookDepthResponse,
